// src/enum/purpose.enums.ts
export enum PreferencePurposeColumn {
  PreferencePurposeName = "preferencePurposeName",
  OrganizationName = "organizationName",
  StandardPreferenceCount = "standardPreferenceCount",
  CreatedBy = "createBy",
  CreatedDate = "createdDate",
  ModifiedBy = "modifiedBy",
  ModifiedDate = "modifiedDate",
}

export enum Sort {
  Ascending = "asc",
  Descending = "desc",
}